+++ CHANGES.rst
@@ -5,7 +5,9 @@ Changelog
 1.3.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fixed problem with ``redirect_url`` not starting with a slash.
+  With ``headerlogin`` as value we would redirect to http://localhost:8080/Ploneheaderlogin.
+  [maurits]
 
 
 1.3.0 (2020-03-26)